refactor: extract DXF document commit context
